one more [long] I missed in the last commit
git-svn-id: https://svn.apache.org/repos/asf/wicket/trunk@996120 13f79535-47bb-0310-9956-ffa450edef68

+++ wicket/src/main/java/org/apache/wicket/protocol/http/RequestUtils.java
@@ -36,7 +36,7 @@ import org.apache.wicket.util.string.Strings;
  */
 public final class RequestUtils
 {
-	public static final long MAX_CACHE_DURATION = 60 * 60 * 24 * 365; // one year, maximum recommended cache duration
+	public static final int MAX_CACHE_DURATION = 60 * 60 * 24 * 365; // one year, maximum recommended cache duration
 
 	/**
 	 * Decode the provided queryString as a series of key/ value pairs and set them in the provided
